pyproject.toml
[project]
name = "hackaton-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiokafka>=0.12.0",
    "datasets>=3.6.0",
    "deep-translator>=1.11.4",
    "httpx-retries>=0.4.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.7.0",
    "torch==2.6",
    "transformers[torch]>=4.52.4",
]
